migrated auth-dbcontext with add-migration init -context name, Authorize works with postman and weather-api
