--- Source/SpacePark-API/Authentication/ApplicationDbContext.cs
@@ -1,58 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore;
-using System.ComponentModel.DataAnnotations;
-
-namespace SpacePark_API.Authentication
-{
-    //https://www.c-sharpcorner.com/article/authentication-and-authorization-in-asp-net-core-web-api-with-json-web-tokens/
-    //https://stackoverflow.com/questions/43767933/entity-framework-core-using-multiple-dbcontexts
-    //use in package manager console:
-    //      "Add-Migration init -Context ApplicationDbContext"
-    //      "Update-Database -Context ApplicationDbContext"
-    public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
-    {
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options)
-        {
-
-        }
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            base.OnModelCreating(builder);
-        }
-    }
-    public class ApplicationUser : IdentityUser
-    {
-    }
-    public static class UserRoles
-    {
-        public const string Admin = "Admin";
-        public const string User = "User";
-    }
-    public class RegisterModel
-    {
-        [Required(ErrorMessage = "User Name is required")]
-        public string Username { get; set; }
-
-        [EmailAddress]
-        [Required(ErrorMessage = "Email is required")]
-        public string Email { get; set; }
-
-        [Required(ErrorMessage = "Password is required")]
-        public string Password { get; set; }
-
-    }
-    public class LoginModel
-    {
-        [Required(ErrorMessage = "User Name is required")]
-        public string Username { get; set; }
-
-        [Required(ErrorMessage = "Password is required")]
-        public string Password { get; set; }
-    }
-    public class Response
-    {
-        public string Status { get; set; }
-        public string Message { get; set; }
-    }
-}
