+++ Source/SpacePark-API/Authentication/ApplicationUser.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace SpacePark_API.Authentication
+{
+    public class ApplicationUser : IdentityUser
+    {
+    }
+}
