+++ Source/SpacePark-API/Authentication/LoginModel.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace SpacePark_API.Authentication
+{
+    public class LoginModel
+    {
+        [Required(ErrorMessage = "User Name is required")]
+        public string Username { get; set; }
+
+        [Required(ErrorMessage = "Password is required")]
+        public string Password { get; set; }
+    }
+}
