+++ Source/SpacePark-API/Authentication/Response.cs
@@ -1,0 +1,8 @@
+﻿namespace SpacePark_API.Authentication
+{
+    public class Response
+    {
+        public string Status { get; set; }
+        public string Message { get; set; }
+    }
+}
