+++ Source/SpacePark-API/Authentication/UserRoles.cs
@@ -1,0 +1,8 @@
+﻿namespace SpacePark_API.Authentication
+{
+    public static class UserRoles
+    {
+        public const string Admin = "Admin";
+        public const string User = "User";
+    }
+}
